zephyr manifest: Use Zephyr from upstream when we don't have patches
It looks better to consume zephyr from the upstream repo when we don't
have patches.

If we want to control when we update the commit, we should use a
commit ID for the revision.  Here I use what was on the OpenAMP fork
for the v3.4-branch.

NOTE: kv260_r5 samples require patches from the v3.5 series so will not
work with this revision setting.

Signed-off-by: Bill Mills <bill.mills@linaro.org>