HACK: Prototype for using upstream open-amp and libmetal
This is a prototype for using open-amp and libmetal code in zephyr builds.
It uses branches of this repo to support the needed zephyr module boiler
plate.

HACKS: wam remote references will go away once this is accepted upstream.

Pros:
* All needed files are provided by just doing the west update
* Don't need to set any ENV vars each time like EXTRA_ZEPHYR_MODULES
* Works with our samples and zephyr stock samples
* The version of open-amp and libmetal are directly controlled in
  the manifest
* It is low maintenance as the module boiler plate is unlikely to change
  much once we are done with the prototyping.

Cons:
* Extra clones of openamp-system-reference
  * Don't let the repo get too big
  * Extra clones have disjoint history so this may reduce clone overhead

I tried a lot of others things and this approach seemed best.

Signed-off-by: Bill Mills <bill.mills@linaro.org>