open-amp-module: get rid of symlink
This technique does not need the symlink and points cmake directly at the
source dir.  We ignore any configured source dir and use the one we setup.

The libmetal INC and LIB statements don't seem to be needed as they point
at the wrong dirs for our manifest setup and it works with out them.

Signed-off-by: Bill Mills <bill.mills@linaro.org>